Fixed wrong printf parameter size on some platforms

+++ examples/cpp/client.cpp
@@ -421,7 +421,7 @@ void ReadSzl_0011_0000()
      if (Check(res,"Read SZL - ID : 0x0011, IDX 0x0000"))
      {
         printf("  LENTHDR : %u\n",(unsigned int)SZL->Header.LENTHDR);
-        printf("  N_DR    : %d\n",SZL->Header.N_DR);
+        printf("  N_DR    : %u\n",(unsigned int)SZL->Header.N_DR);
         printf("Dump (%d bytes) :\n",Size);
         hexdump(&Buffer,Size);
      }
